Introduce hips.utils.testing
